Adds comments on how to use mongoose 'virtual' methods
The virtual getters are for composing values (like first and last
name), without saving the aggregated values, whereas the setters
are used to decompose values and save them as individual values.

The user interface would look like it saves a password, but there
is no password property, instead a hash value is saved to the model.